requirements.txt
streamlit
opencv-python
pillow
pandas
numpy
supabase
python-dotenv
